Migration Fix QuizScore modelBuilder One-Many relations.
Follow new folder structor.

ChatRoomController ad GetAll endpoint
